+++ triggers/http/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @nanoservice-ts/trigger-http
 
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies
+  - @nanoservice-ts/helper@0.1.5
+  - @nanoservice-ts/runner@0.1.20
+  - @nanoservice-ts/if-else@0.0.24
+  - @nanoservice-ts/api-call@0.1.23
+
 ## 0.0.26
 
 ### Patch Changes